+++ test/conftest.py
@@ -1,21 +1,30 @@
-import pytest
-
-from functools import wraps
 from datetime import datetime
+from functools import wraps
+
+import pytest
 from flask import url_for
 from flask_login import current_user
 
-from tracker import create_app, db as flask_db
-from tracker.user import random_string, hash_password
+from tracker import db as flask_db
+from tracker import create_app
 from tracker.advisory import advisory_get_label
-from tracker.model.user import User
-from tracker.model.enum import UserRole, Severity, Remote, Affected, Publication, affected_to_status, highest_severity
 from tracker.model.advisory import Advisory
-from tracker.model.cve import CVE, issue_types
+from tracker.model.cve import CVE
+from tracker.model.cve import issue_types
 from tracker.model.cvegroup import CVEGroup
 from tracker.model.cvegroupentry import CVEGroupEntry
 from tracker.model.cvegrouppackage import CVEGroupPackage
+from tracker.model.enum import Affected
+from tracker.model.enum import Publication
+from tracker.model.enum import Remote
+from tracker.model.enum import Severity
+from tracker.model.enum import UserRole
+from tracker.model.enum import affected_to_status
+from tracker.model.enum import highest_severity
 from tracker.model.package import Package
+from tracker.model.user import User
+from tracker.user import hash_password
+from tracker.user import random_string
 
 DEFAULT_USERNAME = 'cyberwehr12345678'
 ERROR_LOGIN_REQUIRED = 'Please log in to access this page.'
