+++ tracker/cli/run.py
@@ -1,7 +1,12 @@
 from click import option
 from flask.cli import pass_script_info
-from config import set_debug_flag, FLASK_HOST, FLASK_PORT, FLASK_DEBUG
-from . import cli
+
+from config import FLASK_DEBUG
+from config import FLASK_HOST
+from config import FLASK_PORT
+from config import set_debug_flag
+
+from .util import cli
 
 
 @cli.command('run', short_help='Runs a development server.')
