+++ tracker/form/__init__.py
@@ -1,3 +1,3 @@
-from .login import LoginForm
 from .cve import CVEForm
 from .group import GroupForm
+from .login import LoginForm
