+++ tracker/model/__init__.py
@@ -1,7 +1,7 @@
+from .advisory import Advisory
 from .cve import CVE
 from .cvegroup import CVEGroup
 from .cvegroupentry import CVEGroupEntry
 from .cvegrouppackage import CVEGroupPackage
-from .advisory import Advisory
 from .package import Package
 from .user import User
