+++ tracker/view/__init__.py
@@ -1,14 +1,13 @@
 from .add import *
-from .edit import *
-from .show import *
-from .delete import *
+from .admin import *
+from .advisory import *
 from .copy import *
-
+from .delete import *
+from .edit import *
+from .error import *
 from .index import *
 from .login import *
-from .advisory import *
-from .todo import *
-from .error import *
-from .admin import *
-from .user import *
+from .show import *
 from .stats import *
+from .todo import *
+from .user import *
